+++ src/Storage/Storage.Management/Models/PSStorageAccount.cs
@@ -72,6 +72,7 @@ namespace Microsoft.Azure.Commands.Management.Storage.Models
             this.EnableSftp = storageAccount.IsSftpEnabled;
             this.EnableLocalUser = storageAccount.IsLocalUserEnabled;
             this.AllowedCopyScope = storageAccount.AllowedCopyScope;
+            this.DnsEndpointType = storageAccount.DnsEndpointType;
         }
         public bool? AllowCrossTenantReplication { get; set; }
 
@@ -163,6 +164,7 @@ namespace Microsoft.Azure.Commands.Management.Storage.Models
 
         public PSImmutableStorageAccount ImmutableStorageWithVersioning { get; set; }
         public string AllowedCopyScope { get; set; }
+        public string DnsEndpointType { get; set; }
 
         public static PSStorageAccount Create(StorageModels.StorageAccount storageAccount, IStorageManagementClient client)
         {